backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pandas>=3.0.0
numpy>=2.2.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
httpx>=0.28.0
python-dotenv>=1.0.1
scipy>=1.15.0
